Feature: Mark the best answer.

+++ database/migrations/2019_11_21_163656_create_answers_table.php
@@ -15,10 +15,10 @@ class CreateAnswersTable extends Migration
     {
         Schema::create('answers', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('question_id');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('question_id');
+            $table->unsignedBigInteger('user_id');
             $table->text('body');
-            $table->integer('votes_count')->default(0);
+            $table->BigInteger('votes_count')->default(0);
             $table->timestamps();
         });
     }
